requirements.txt
asgiref==3.8.1
Django==5.1
django-allauth==64.2.0
django-cleanup==8.1.0
django-colorfield==0.11.0
django-environ==0.11.2
django-htmx==1.19.0
django-tenants==3.7.0
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
sqlparse==0.5.1
whitenoise==6.7.0
